+++ src/guard/mod.rs
@@ -1,6 +1,7 @@
 use std::io::Write;
 
 use crate::model::error::guard_errors::AuthError;
+use crate::model::guard::auth::ValidateResult;
 use base64::decode;
 use rocket::async_trait;
 use rocket::http::Status;
@@ -8,15 +9,8 @@ use rocket::request::{FromRequest, Outcome};
 use rocket::Request;
 use sha2::{Digest, Sha256};
 
-use crate::repository::metadata_repository::CheckAuthResult;
+use crate::model::service::metadata::CheckAuthResult;
 use crate::service::api_service;
-
-/// used to represent the result of calling `Auth::validate`
-pub enum ValidateResult {
-    Ok,
-    NoPasswordSet,
-    Invalid,
-}
 
 #[derive(Debug)]
 pub struct Auth {
