+++ src/model/guard/auth.rs
@@ -1,0 +1,6 @@
+/// used to represent the result of calling `Auth::validate`
+pub enum ValidateResult {
+    Ok,
+    NoPasswordSet,
+    Invalid,
+}
