+++ src/model/mod.rs
@@ -1,4 +1,6 @@
-pub mod db;
 pub mod error;
+pub mod guard;
+pub mod repository;
 pub mod request;
 pub mod response;
+pub mod service;
