+++ src/model/response/folder_responses.rs
@@ -1,4 +1,4 @@
-use crate::model::db::{FileRecord, Folder};
+use crate::model::repository::{FileRecord, Folder};
 use crate::model::response::file_responses::FileMetadataResponse;
 use crate::model::response::BasicMessage;
 use rocket::serde::{json::Json, Serialize};
