+++ src/repository/folder_repository.rs
@@ -1,10 +1,10 @@
-use crate::model::db;
+use crate::model::repository;
 use rusqlite::{params, Connection, Row, Rows};
 
-pub fn get_by_id(id: Option<u32>, con: &Connection) -> Result<db::Folder, rusqlite::Error> {
+pub fn get_by_id(id: Option<u32>, con: &Connection) -> Result<repository::Folder, rusqlite::Error> {
     // if id is none, we're talking about the root folder
     if id.is_none() {
-        return Ok(db::Folder {
+        return Ok(repository::Folder {
             id: Some(0), // will never collide with an id since sqlite starts with 1
             name: String::from("root"),
             parent_id: None,
@@ -17,7 +17,7 @@ pub fn get_by_id(id: Option<u32>, con: &Connection) -> Result<db::Folder, rusqli
         .unwrap();
 
     let row_mapper = |row: &Row| {
-        Ok(db::Folder {
+        Ok(repository::Folder {
             id: Some(row.get(0)?),
             name: row.get(1)?,
             parent_id: row.get(2).ok().or_else(|| None),
@@ -33,7 +33,7 @@ pub fn get_by_id(id: Option<u32>, con: &Connection) -> Result<db::Folder, rusqli
 pub fn get_child_folders(
     id: Option<u32>,
     con: &Connection,
-) -> Result<Vec<db::Folder>, rusqlite::Error> {
+) -> Result<Vec<repository::Folder>, rusqlite::Error> {
     let mut pst = if id.is_some() {
         con.prepare(include_str!(
             "../assets/queries/folder/get_child_folders_with_id.sql"
@@ -45,7 +45,7 @@ pub fn get_child_folders(
         ))
         .unwrap()
     };
-    let mut folders = Vec::<db::Folder>::new();
+    let mut folders = Vec::<repository::Folder>::new();
     let mut rows: Rows;
     if id.is_some() {
         rows = pst.query([id.unwrap()])?;
@@ -54,7 +54,7 @@ pub fn get_child_folders(
     }
     while let Some(row) = rows.next()? {
         // these folders are guaranteed to have a parent folder id
-        folders.push(db::Folder {
+        folders.push(repository::Folder {
             id: Some(row.get(0)?),
             name: row.get(1)?,
             parent_id: row.get(2)?,
@@ -66,14 +66,17 @@ pub fn get_child_folders(
 /// creates a folder record in the database.
 /// This does not do any checks on folder parent id or any other data,
 /// and that must be done before this function is called
-pub fn create_folder(folder: &db::Folder, con: &Connection) -> Result<db::Folder, rusqlite::Error> {
+pub fn create_folder(
+    folder: &repository::Folder,
+    con: &Connection,
+) -> Result<repository::Folder, rusqlite::Error> {
     let mut pst = con
         .prepare(include_str!("../assets/queries/folder/create_folder.sql"))
         .unwrap();
     return match folder.parent_id {
         Some(id) => {
             let folder_id = pst.insert(rusqlite::params![folder.name, id])? as u32;
-            Ok(db::Folder {
+            Ok(repository::Folder {
                 id: Some(folder_id),
                 name: String::from(&folder.name),
                 parent_id: folder.parent_id,
@@ -82,7 +85,7 @@ pub fn create_folder(folder: &db::Folder, con: &Connection) -> Result<db::Folder
         None => {
             let folder_id =
                 pst.insert(rusqlite::params![folder.name, rusqlite::types::Null])? as u32;
-            Ok(db::Folder {
+            Ok(repository::Folder {
                 id: Some(folder_id),
                 name: String::from(&folder.name),
                 parent_id: folder.parent_id,
@@ -94,7 +97,7 @@ pub fn create_folder(folder: &db::Folder, con: &Connection) -> Result<db::Folder
 /// updates a folder record in the database.
 /// This does not perform any checks on folder info, and that must be done
 /// before this function is called
-pub fn update_folder(folder: &db::Folder, con: &Connection) -> Result<(), rusqlite::Error> {
+pub fn update_folder(folder: &repository::Folder, con: &Connection) -> Result<(), rusqlite::Error> {
     let mut pst = con
         .prepare(include_str!("../assets/queries/folder/update_folder.sql"))
         .unwrap();
@@ -110,7 +113,7 @@ pub fn update_folder(folder: &db::Folder, con: &Connection) -> Result<(), rusqli
 pub fn get_files_for_folder(
     id: Option<u32>,
     con: &Connection,
-) -> Result<Vec<db::FileRecord>, rusqlite::Error> {
+) -> Result<Vec<repository::FileRecord>, rusqlite::Error> {
     let mut pst = if id.is_some() {
         con.prepare(include_str!(
             "../assets/queries/folder_file/get_child_files_with_id.sql"
@@ -123,7 +126,7 @@ pub fn get_files_for_folder(
         .unwrap()
     };
     let row_mapper = |row: &Row| {
-        Ok(db::FileRecord {
+        Ok(repository::FileRecord {
             id: Some(row.get(0)?),
             name: row.get(1)?,
             hash: row.get(2)?,
@@ -135,7 +138,7 @@ pub fn get_files_for_folder(
         pst.query_map([], row_mapper)?
     };
 
-    let mut files: Vec<db::FileRecord> = Vec::new();
+    let mut files: Vec<repository::FileRecord> = Vec::new();
     for file in mapped.into_iter() {
         files.push(file?);
     }
