+++ src/repository/metadata_repository.rs
@@ -1,6 +1,7 @@
 use rusqlite::Connection;
 
 use crate::guard::Auth;
+use crate::model::service::metadata::CheckAuthResult;
 
 /// returns the current version of the database as a String
 pub fn get_version(con: &mut Connection) -> Result<String, rusqlite::Error> {
@@ -10,18 +11,6 @@ pub fn get_version(con: &mut Connection) -> Result<String, rusqlite::Error> {
         |row| row.get(0),
     );
     return result;
-}
-
-/// represents the result of comparing a password to the database value
-pub enum CheckAuthResult {
-    /// The passed authorization matches what's in the database
-    Valid,
-    /// The passed authorization does not match what's in the database
-    Invalid,
-    /// there is no auth field in the database, and one needs to be set
-    Missing,
-    /// The database encountered an error trying to retrieve authorization
-    DbError,
 }
 
 /// retrieves the encrypted authentication string for requests in the database
